panda_data_hub/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
aiohttp==3.9.1
pyyaml==6.0.1
rqdatac>=2.9.0
tqsdk>=3.2.10
tushare>=1.4.21
pandas_market_calendars>=5.1.0
chinese_calendar>=1.10.0
xtquant